requirements.txt
playwright
python-dotenv


